pyproject.toml
[tool.poetry]
name = "ten-second-tools"
version = "0.1.0"
description = ""
authors = ["armckinney"]
packages = [
    { include = "Psyncronous", from = "src" },
    { include = "Rando", from = "src" },
    { include = "tax", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.2"
networkx = "^2.7.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
flake8 = "^4.0.1"
pep8-naming = "^0.12.1"
black = "^22.1.0"
mypy = "^0.931"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
